08.12 - final refactorings (step 1 - refactor the users feature) - create a users package. move all the user related code to the users package
