+++ src/main/java/com/jparkkennaby/store/payments/CheckoutService.java
@@ -3,12 +3,12 @@ package com.jparkkennaby.store.payments;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.jparkkennaby.store.services.AuthService;
 import com.jparkkennaby.store.entities.Order;
 import com.jparkkennaby.store.exceptions.CartEmptyException;
 import com.jparkkennaby.store.exceptions.CartNotFoundException;
 import com.jparkkennaby.store.repositories.CartRepository;
 import com.jparkkennaby.store.repositories.OrderRepository;
-import com.jparkkennaby.store.services.AuthService;
 import com.jparkkennaby.store.services.CartService;
 
 import lombok.RequiredArgsConstructor;
