+++ src/main/java/com/jparkkennaby/store/security/ProductSecurityRules.java
@@ -5,7 +5,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configurers.AuthorizeHttpRequestsConfigurer;
 import org.springframework.stereotype.Component;
 
-import com.jparkkennaby.store.entities.Role;
+import com.jparkkennaby.store.users.Role;
 
 @Component
 public class ProductSecurityRules implements SecurityRules {
