+++ src/main/java/com/jparkkennaby/store/services/JwtService.java
@@ -5,7 +5,7 @@ import java.util.Date;
 import org.springframework.stereotype.Service;
 
 import com.jparkkennaby.store.config.JwtConfig;
-import com.jparkkennaby.store.entities.User;
+import com.jparkkennaby.store.users.User;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
