+++ src/main/java/com/jparkkennaby/store/services/UserDetailsServiceImpl.java
@@ -8,7 +8,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.jparkkennaby.store.repositories.UserRepository;
+import com.jparkkennaby.store.users.UserRepository;
 
 import lombok.AllArgsConstructor;
