+++ src/main/java/com/jparkkennaby/store/users/ChangePasswordRequest.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.dtos;
+package com.jparkkennaby.store.users;
 
 import lombok.Data;
