+++ src/main/java/com/jparkkennaby/store/users/ProfileRepository.java
@@ -1,8 +1,6 @@
-package com.jparkkennaby.store.repositories;
+package com.jparkkennaby.store.users;
 
 import org.springframework.data.repository.CrudRepository;
-
-import com.jparkkennaby.store.entities.Profile;
 
 public interface ProfileRepository extends CrudRepository<Profile, Long> {
 }
