+++ src/main/java/com/jparkkennaby/store/users/UserDto.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.dtos;
+package com.jparkkennaby.store.users;
 
 // import java.time.LocalDateTime;
 // import com.fasterxml.jackson.annotation.JsonFormat;
@@ -12,7 +12,8 @@ import lombok.Getter;
 /**
  * Data Transfer Objets (DTO)
  * 
- * DTO's are used to manage the fields we want to expose in our public facing REST API
+ * DTO's are used to manage the fields we want to expose in our public facing
+ * REST API
  */
 
 @Getter
@@ -24,9 +25,10 @@ public class UserDto {
     private String name;
     private String email;
 
-    // @JsonInclude(JsonInclude.Include.NON_NULL) // exlcude null fields from response
+    // @JsonInclude(JsonInclude.Include.NON_NULL) // exlcude null fields from
+    // response
     // private String phoneNumber;
 
     // @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    // private LocalDateTime createdAt; // filed populated in the UserMapper class 
+    // private LocalDateTime createdAt; // filed populated in the UserMapper class
 }
