+++ src/main/java/com/jparkkennaby/store/users/UserRepository.java
@@ -1,17 +1,17 @@
-package com.jparkkennaby.store.repositories;
+package com.jparkkennaby.store.users;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.jparkkennaby.store.entities.User;
-
 /**
  * UserRepository:
  * 
- * - Uses a JpaRepository instead of CrudRepository so that the findAll method returns
+ * - Uses a JpaRepository instead of CrudRepository so that the findAll method
+ * returns
  * a List, instead of an Iterable object.
- * - Returning a List allows us to use the stream api to map User objects to the UserDto object.
+ * - Returning a List allows us to use the stream api to map User objects to the
+ * UserDto object.
  */
 
 public interface UserRepository extends JpaRepository<User, Long> {
